Move middleware package from interfaces layer to infrastructure layer
