+++ user-rest-service/infrastructure/middleware/auth.go
